UnfabricatableError should include the message of the original exception
When custom #const_missing is defined and it is possible to get
NameError/NoMethodError from inside #const_missing. This is very
important since widely-used ActiveSupport::Dependencies's
"autoload-like" feature is made possible using #const_missing and
there's a lot of things that will whappen when it's called.

Elevate the original message so we can easily understand what's going on
in #const_missing.

(For more about Module#const_missing, see:
http://ruby-doc.org/core-2.1.1/Module.html#method-i-const_missing)