Do not include Rake::DSL in the top-level scope
fixes #289.

The commit fd0424bf changed the `lib/fabrication.rb` file to include
the `Rake::DSL` module, but it turns out that this call is unnecessary
since Fabrication's rake definition is in fact compatible with both of
Rake 0.9.* and 10.*. Also, it had brought the deprecated feature back
in without showing any deprecation warnings.

This commit removes the include call since it is 100% safe to do so.