Allow overrides to be `nil`
This fixes a problem where attribute methods would be treated as missing when their override value was set to `nil`.

Instead of relying on the value of the attribute, this commit changes the behavior to use `fetch` on the attributes hash. Only a truly missing attribute will now trigger the `super` call in `method_missing`.